requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
DateTime==4.9
Django==4.1.4
future==0.18.2
idna==3.4
peewee==3.15.4
pefile==2021.9.3
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
telebot==0.0.4
typish==1.9.3
ujson==5.6.0
urllib3==1.26.13
zope.interface==5.5.2
